IMixer | new Mixer implementation (#240)
* [WIP] new mixer

* [WIP] new mixer

* [WIP] new mixer

* [WIP] new mixer

* [WIP] new mixer

* code rev

* [WIP] PCMDriver using new mixer

* fix build

* fix tests

* code rev

* clean left-over

* fix test

* fix test

* code rev

* refactor

* code rev

* code rev

* code rev#

* code rev

* code rev

* code review sonarcloud

+++ sdl2-hyper-sonic-drivers/src/HyperSonicDrivers/audio/converters/SimpleRateConverter.hpp
@@ -30,9 +30,9 @@ namespace HyperSonicDrivers::audio::converters
 
     public:
         SimpleRateConverter(uint32_t inrate, uint32_t outrate);
-        int flow(IAudioStream& input, int16_t* obuf, uint32_t osamp, const uint16_t vol_l, const uint16_t vol_r) override;
+        size_t flow(IAudioStream& input, int16_t* obuf, uint32_t osamp, const uint16_t vol_l, const uint16_t vol_r) override;
         
-        int drain(int16_t* obuf, uint32_t osamp, const uint16_t vol) override
+        size_t drain(int16_t* obuf, uint32_t osamp, const uint16_t vol) override
         {
             return 0;
         }
@@ -67,7 +67,7 @@ namespace HyperSonicDrivers::audio::converters
      * Return number of sample pairs processed.
      */
     template<bool stereo, bool reverseStereo>
-    int SimpleRateConverter<stereo, reverseStereo>::flow(IAudioStream& input, int16_t* obuf, uint32_t osamp, const uint16_t vol_l, const uint16_t vol_r)
+    size_t SimpleRateConverter<stereo, reverseStereo>::flow(IAudioStream& input, int16_t* obuf, uint32_t osamp, const uint16_t vol_l, const uint16_t vol_r)
     {
         const int16_t* ostart = obuf;
         const int16_t* oend = obuf + osamp * 2;
